feat: Add FileService.swift for importing and exporting files
This commit adds the `FileService.swift` file, which contains the implementation for importing and exporting files. The `importFiles(at:)` method allows multiple files to be imported at once, while the `exportFiles(at:)` method exports multiple files. Additionally, the `importFile(at:)` and `exportFile(to:)` methods allow individual files to be imported or exported respectively.

The commit also includes a new struct called `DirectoryStructure`, which represents the desired directory structure for imported files. It has three boolean properties that determine whether the files should be separated by platform, region, or alphabetically.

Lastly, this commit adds a new enum case to the existing `Platform` enum called `.gb`. This case represents Game Boy games and is used in various parts of the codebase.